requirements.txt
minigram-py==0.3.0
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
